تعديل الدفع واضافه المحفظه والتعديل عليهما بشكل ممتاز

+++ src/main/java/org/example/tears/OutDTO/PreviewResponseDto.java
@@ -4,6 +4,6 @@ import lombok.Data;
 
 @Data
 public class PreviewResponseDto {
-    private Integer estimatedPrice;
+    private Double estimatedPrice;
     private String details;
 }
